Create app.js
So this was my first intro to the concept of routes, none of it is broken up into smaller segments. All of the routes are all there which makes it look sloppy but it does work fin. I've been working on another similar project and to break up the routes to smaller pieces it has its own drawback of a bunch of "requirements" per new broken segment.